SM-1 allow ip6 field and multiple address for ip4/6
This is a squashed and reworked commit; original commit messages are
below.
Also, overlay/generic/lib/svc/method/net-physical doesn't handle
multiple ipv4 addresses like it did in the older versions. -Nahum

This commit will allow you to specify an ip6 field in a nic
ex: "ip6" = "1::a302/12"
it will also allow multiple ips and stroke notation in ipv4 and ipv6
ex: "ip" = "12.3.4.5/16,3.4.5.6/32"
it will accept spaces, and if there is no stroke, it will use the
gateway field just like it used to.

this relies on an updated net-physical to be automatic, and
there is a commit for that

ip6 modifications to joyent-minimal brand

this is separate in case joyent doesn't want to pull :)

no anti-spoofing when ip6 set to autoconfig

no anti-spoofing when ip6 set to auto for joyent-minimal

fixed reset error

Conflicts:

	overlay/generic/usr/lib/brand/joyent-minimal/statechange
	overlay/generic/usr/lib/brand/joyent/statechange
(cherry picked from commit deab8d3a65a350dfb1480c0ecd7529b56aa9e920)